Fix decoder return convention and payload size plumbing

+++ runtime/lib/transport/gpu_verbs/coprocessor/coprocessor_functions/echo_decoder.hpp
@@ -16,6 +16,10 @@
 #include <cstdint>
 namespace catalyst::transport::gpu_verbs {
 
-__device__ inline std::uint64_t echo_decode(std::uint64_t syndrome) { return syndrome; }
+// Passthrough self-test: returns the word unchanged.
+__device__ inline std::int64_t echo_decode(std::uint64_t syndrome)
+{
+    return static_cast<std::int64_t>(syndrome);
+}
 
 } // namespace catalyst::transport::gpu_verbs
